chore(client): Rename a file
